后端商品

+++ ShoppingMall/src/main/java/com/zby/demo/service/ProductService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface ProductService {
     public List<Product> getAll();
     public boolean updateProduct(Product product);
-    public boolean delProduct(Long id);
+    public boolean delProduct(int id);
 }
